part 7 done. left to do: english changes

+++ Alternative banking system/DTO/src/objects/customers/loanInfo/ActiveLoanInfoDTO.java
@@ -4,7 +4,7 @@ public class ActiveLoanInfoDTO extends LoanInfoDTO{
     private int nextTimePayment;
     private double expectedPayment;
 
-    public ActiveLoanInfoDTO(String loanName, String loanCategory, double sizeNoInterest, int interestPerPayment, int timePerPayment, String status, int nextTimePayment, double expectedPayment) {
+    public ActiveLoanInfoDTO(String loanName, String loanCategory, double sizeNoInterest, double interestPerPayment, int timePerPayment, String status, int nextTimePayment, double expectedPayment) {
         super(loanName, loanCategory, sizeNoInterest, interestPerPayment, timePerPayment, status);
             this.nextTimePayment = nextTimePayment;
             this.expectedPayment = expectedPayment;
