+++ Alternative banking system/DTO/src/objects/loans/PendingLoanDTO.java
@@ -13,7 +13,7 @@ public class PendingLoanDTO extends  NewLoanDTO{
 
 
 
-    public PendingLoanDTO(String loanID, String borrowerName, String loanCategory, double sizeNoInterest, int timeLimitOfLoan, int interestPerPayment, int timePerPayment, String status, Map<String, Double> listOfLenders, Double collectedSoFar, Double sumLeftToBeCollected) {
+    public PendingLoanDTO(String loanID, String borrowerName, String loanCategory, double sizeNoInterest, int timeLimitOfLoan, double interestPerPayment, int timePerPayment, String status, Map<String, Double> listOfLenders, Double collectedSoFar, Double sumLeftToBeCollected) {
         super(loanID, borrowerName, loanCategory, sizeNoInterest, timeLimitOfLoan, interestPerPayment, timePerPayment, status);
         this.listOfLenders = listOfLenders;
         this.collectedSoFar = collectedSoFar;
